+++ course-signup/impl/src/test/java/uk/ac/ox/oucs/vle/XcriOxCapPopulatorImplTest.java
@@ -9,7 +9,7 @@ public class XcriOxCapPopulatorImplTest {
 	@Test
 	public void testParse() {
 		String original = "http://www.lsidtc.ox.ac.uk/the-course/core-modules\r\n\r\nFormal Assessment:";
-		String output = XcriOxCapPopulatorImpl.parse(original);
+		String output = XcriOxCapPopulator.parse(original);
 		assertTrue(output.contains("href=\"http://www.lsidtc.ox.ac.uk/the-course/core-modules\""));
 	}
 }
